Add functionality create new place (room)

+++ app/src/main/java/ie/setu/travelnotes/data/room/repositories/PlaceRepository.kt
@@ -3,18 +3,26 @@ package ie.setu.travelnotes.data.room.repositories
 import ie.setu.travelnotes.data.PlaceModel
 import ie.setu.travelnotes.data.room.PlaceDAO
 import kotlinx.coroutines.flow.Flow
+import timber.log.Timber
 import javax.inject.Inject
 
 class PlaceRepository @Inject
 constructor(private val placeDAO: PlaceDAO) {
+
     fun getAll(): Flow<List<PlaceModel>> = placeDAO.getAll()
 
-    suspend fun insert(place: PlaceModel)
-    { placeDAO.insert(place) }
+    fun getAllByUser(userId: String): Flow<List<PlaceModel>> = placeDAO.getAllByUser(userId)
 
-    suspend fun update(place: PlaceModel)
-    { placeDAO.update(place) }
+    suspend fun insert(place: PlaceModel, userId: String) {
+        place.userId = userId
+        placeDAO.insert(place)
+    }
 
-    suspend fun delete(place: PlaceModel)
-    { placeDAO.delete(place) }
+    suspend fun update(place: PlaceModel, userId: String) {
+        if (place.userId == userId) { placeDAO.update(place) }
+    }
+
+    suspend fun delete(place: PlaceModel, userId: String) {
+        if (place.userId == userId) { placeDAO.delete(place) }
+    }
 }
